requirements.txt
torch==1.5.1+cpu
numpy==1.16.5
flask==1.1.1
gunicorn
matplotlib==3.3.1
pillow==6.2.0